requirements.txt
# Core
pydantic>=2.0,<3.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
pdfplumber>=0.10,<1.0
pyyaml>=6.0,<7.0

# CLI & Display
rich>=13.0,<14.0

# MITRE ATT&CK
stix2>=3.0,<4.0

# Testing
pytest>=7.0,<9.0
pytest-cov>=4.0,<6.0
